+++ HISTORY.rst
@@ -14,6 +14,8 @@ Development
 * Decode Content-Type: text/plain
 * Default key web.Application is "web" instead of "app" (#88)
 * Fix default encode/decode charset
+* Add cpus factor for processes
+